HSEARCH-3791 Move ElasticsearchTypeNameMappingTestUtils to the util module and rename it
So that we can add more utils related to index metadata to it (see the next commits).
