+++ util/internal/integrationtest/backend/elasticsearch/src/main/java/org/hibernate/search/util/impl/integrationtest/backend/elasticsearch/ElasticsearchIndexMetadataTestUtils.java
@@ -4,13 +4,13 @@
  * License: GNU Lesser General Public License (LGPL), version 2.1 or later
  * See the lgpl.txt file in the root directory or <http://www.gnu.org/licenses/lgpl-2.1.html>.
  */
-package org.hibernate.search.integrationtest.backend.elasticsearch.mapping;
+package org.hibernate.search.util.impl.integrationtest.backend.elasticsearch;
 
 import com.google.gson.JsonObject;
 
-public class ElasticsearchTypeNameMappingTestUtils {
+public class ElasticsearchIndexMetadataTestUtils {
 
-	private ElasticsearchTypeNameMappingTestUtils() {
+	private ElasticsearchIndexMetadataTestUtils() {
 	}
 
 	public static JsonObject mappingWithoutAnyProperty() {
